#20 - Improves video streaming service.
By using a more polling-like service we can ask the server if there are any more segments available since the last one, rather than requesting the newest all the time. This reduces bandwidth and also avoids jumps ahead in the stream when the browser took too long to load a segment and the expected next one is no longer the newest.